SmallMolecules and docking viewers centralized in SmallMoleculesViewer
