+++ pwchem/viewers/__init__.py
@@ -5,4 +5,4 @@ from .viewers_data import BioinformaticsDataViewer
 from .viewer_pockets import *
 from .viewer_protocol_consensus_pockets import ViewerConsensusPockets
 from .viewer_smallMols import SmallMoleculesViewer
-from .viewer_docking import DockingViewer, ProtConsensusDockingViewer
+from .viewer_docking import ProtConsensusDockingViewer
